android: Restore non-fullscreen "fix"; default to 0ms warmup
If we warmup *at all*, this "fix" breaks down. This may be related to
Unity's splash screen forcing fullscreen mode for some inane reason.

It'd be great if we could render behind the system bars at some point
but the hacks to prevent this thing from forcing fullscreen might make
that difficult, which is frustrating....